% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{gemcox-package}
\alias{gemcox-package}
\title{gemcox: Gaussian mixture of Cox models for mechanism subgroups}
\description{
Fits latent subgroups whose survival mechanisms (Cox coefficient
vectors) differ, with a shared baseline hazard. The main entry points
are \code{\link[=gemcox]{gemcox()}}, \code{\link[=predict.gemcox]{predict.gemcox()}} and \code{\link[=gemcox_heterogeneity_test]{gemcox_heterogeneity_test()}}.
}
\author{
\strong{Maintainer}: Ziwei Zhao \email{zhaoziwe@gmail.com}

Authors:
\itemize{
  \item Ziwei Zhao \email{zhaoziwe@gmail.com}
}

}
\keyword{internal}
